requirements.txt
python-dotenv==1.1.0
langchain==0.3.25
langchain-anthropic==0.3.12
